Break out 'up' code for running a workload
