+++ crates/controller/src/scheduler.rs
@@ -1,13 +1,9 @@
 use std::{collections::HashMap, sync::{Arc, RwLock}, path::{Path, PathBuf}};
 
 use anyhow::{bail, Context};
-use spin_http_engine::{HttpTrigger, TlsConfig, HttpTriggerExecutionConfig};
-use spin_manifest::ApplicationTrigger;
-use spin_redis_engine::RedisTrigger;
-use spin_trigger::{ExecutionOptions, run_trigger};
 use tempfile::TempDir;
 
-use crate::{schema::{WorkloadId, WorkloadManifest}, store::WorkStore, WorkloadSpec, WorkloadEvent};
+use crate::{schema::WorkloadId, store::WorkStore, WorkloadSpec, WorkloadEvent, run::run};
 
 #[async_trait::async_trait]
 pub(crate) trait Scheduler {
@@ -30,12 +26,12 @@ impl LocalScheduler {
     }
 }
 
-struct RunningWorkload {
-    work_dir: WorkingDirectory,
-    handle: RunHandle,
+pub(crate) struct RunningWorkload {
+    pub(crate) work_dir: WorkingDirectory,
+    pub(crate) handle: RunHandle,
 }
 
-enum RunHandle {
+pub(crate) enum RunHandle {
     Fut(tokio::task::JoinHandle<()>),
 }
 
@@ -63,103 +59,10 @@ impl LocalScheduler {
         // Identify the application type
         // Instantiate the relevant trigger
         // Start the relevant trigger
-        let running = self.start_workload_from_spec(workload, spec).await?;
+        let running = run(workload, spec, &self.notification_sender).await?;
         // Stash the task
         self.running.write().unwrap().insert(workload.clone(), running);
         Ok(())
-    }
-
-    async fn start_workload_from_spec(&self, workload: &WorkloadId, spec: WorkloadSpec) -> anyhow::Result<RunningWorkload> {
-        let working_dir_holder = match &spec.opts.tmp {
-            None => WorkingDirectory::Temporary(tempfile::tempdir()?),
-            Some(d) => WorkingDirectory::Given(d.to_owned()),
-        };
-        let working_dir = working_dir_holder.path();
-
-        let mut app = match &spec.manifest {
-            WorkloadManifest::File(manifest_file) => {
-                let bindle_connection = spec.opts.bindle_connection();
-                spin_loader::from_file(manifest_file, working_dir, &bindle_connection).await?
-            },
-            WorkloadManifest::Bindle(bindle) => match &spec.opts.server {
-                Some(server) => spin_loader::from_bindle(bindle, server, working_dir).await?,
-                _ => bail!("Loading from a bindle requires a Bindle server URL"),
-            },
-        };
-        append_env(&mut app, &spec.opts.env)?;
-
-        if let Some(ref mut resolver) = app.config_resolver {
-            // TODO(lann): This should be safe but ideally this get_mut would be refactored away.
-            let resolver = Arc::get_mut(resolver)
-                .context("Internal error: app.config_resolver unexpectedly shared")?;
-            // TODO(lann): Make config provider(s) configurable.
-            resolver.add_provider(spin_config::provider::env::EnvProvider::default());
-        }
-
-        let tls = match (spec.opts.tls_key.clone(), spec.opts.tls_cert.clone()) {
-            (Some(key_path), Some(cert_path)) => {
-                if !cert_path.is_file() {
-                    bail!("TLS certificate file does not exist or is not a file")
-                }
-                if !key_path.is_file() {
-                    bail!("TLS key file does not exist or is not a file")
-                }
-                Some(TlsConfig {
-                    cert_path,
-                    key_path,
-                })
-            }
-            (None, None) => None,
-            _ => unreachable!(),
-        };
-
-        let wasmtime_config = spec.opts.wasmtime_default_config()?;
-
-        let follow = spec.opts.follow_components();
-
-        let tx = self.notification_sender.clone();
-        let id = workload.clone();
-
-        let jh = match &app.info.trigger {
-            ApplicationTrigger::Http(_) => {
-                tokio::spawn(async move {
-                    let r = run_trigger(
-                        app,
-                        ExecutionOptions::<HttpTrigger>::new(
-                            spec.opts.log.clone(),
-                            follow,
-                            HttpTriggerExecutionConfig::new(spec.opts.address, tls),
-                        ),
-                        Some(wasmtime_config),
-                    ).await;
-                    let err = match r {
-                        Ok(()) => None,
-                        Err(e) => Some(Arc::new(e)),
-                    };
-                    // TODO: this should update the workflow status in the scheduler's record
-                    let _ = tx.send(WorkloadEvent::Stopped(id, err));
-                })
-            }
-            ApplicationTrigger::Redis(_) => {
-                tokio::spawn(async move {
-                    let r = run_trigger(
-                        app,
-                        ExecutionOptions::<RedisTrigger>::new(spec.opts.log.clone(), follow, ()),
-                        Some(wasmtime_config),
-                    ).await;
-                    let err = match r {
-                        Ok(()) => None,
-                        Err(e) => Some(Arc::new(e)),
-                    };
-                    let _ = tx.send(WorkloadEvent::Stopped(id, err));
-                })
-            }
-        };
-
-        Ok(RunningWorkload {
-            work_dir: working_dir_holder,
-            handle: RunHandle::Fut(jh),
-        })
     }
 
     async fn restart_workload(&self, workload: &WorkloadId, spec: WorkloadSpec, current: RunningWorkload) -> anyhow::Result<()> {
@@ -182,25 +85,16 @@ impl RunningWorkload {
     }
 }
 
-enum WorkingDirectory {
+pub(crate) enum WorkingDirectory {
     Given(PathBuf),
     Temporary(TempDir),
 }
 
 impl WorkingDirectory {
-    fn path(&self) -> &Path {
+    pub(crate) fn path(&self) -> &Path {
         match self {
             Self::Given(p) => p,
             Self::Temporary(t) => t.path(),
         }
     }
-}
-
-fn append_env(app: &mut spin_manifest::Application, env: &[(String, String)]) -> anyhow::Result<()> {
-    for c in app.components.iter_mut() {
-        for (k, v) in env {
-            c.wasm.environment.insert(k.clone(), v.clone());
-        }
-    }
-    Ok(())
 }
